requirements.txt
streamlit
PyPDF2
streamlit-extras
openai
pandas
streamlit-aggrid
python-dotenv